[Feature and frontend] vacation confirmation and offer winner

+++ pharmacy/src/main/java/com/isa/pharmacy/service/OrderService.java
@@ -1,20 +1,37 @@
 package com.isa.pharmacy.service;
 
+import com.isa.pharmacy.controller.dto.SupplierOfferDto;
+import com.isa.pharmacy.controller.exception.InvalidActionException;
 import com.isa.pharmacy.controller.exception.NotFoundException;
-import com.isa.pharmacy.domain.Order;
-import com.isa.pharmacy.domain.SupplierOffer;
+import com.isa.pharmacy.controller.mapping.SupplierOfferMapper;
+import com.isa.pharmacy.domain.*;
 import com.isa.pharmacy.repository.OrderRepository;
+import com.isa.pharmacy.service.interfaces.IEmailService;
+import com.isa.pharmacy.service.interfaces.IMedicinePharmacyService;
 import com.isa.pharmacy.service.interfaces.IOrderService;
+import com.isa.pharmacy.service.interfaces.ISupplierOfferService;
+import com.isa.pharmacy.users.domain.PharmacyAdmin;
+import com.isa.pharmacy.users.service.interfaces.IPharmacyAdminService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @Service
 public class OrderService implements IOrderService {
     @Autowired
     private OrderRepository orderRepository;
+    @Autowired
+    private ISupplierOfferService supplierOfferService;
+    @Autowired
+    private IEmailService emailService;
+    @Autowired
+    private IMedicinePharmacyService medicinePharmacyService;
+    @Autowired
+    private IPharmacyAdminService pharmacyAdminService;
 
     public List<Order> getAll(){
         List<Order> orders= orderRepository.findAll();
@@ -31,6 +48,9 @@ public class OrderService implements IOrderService {
         Order order = orderRepository.findOrderById(id);
         if(order == null)
             throw new NotFoundException("Order has already been deleted");
+        List<SupplierOfferDto> orderOfferList = supplierOfferService.offersByOrderId(order.getId());
+        if(!orderOfferList.isEmpty())
+            throw new InvalidActionException("You can't delete order with offers");
         orderRepository.delete(order);
     }
 
@@ -52,5 +72,52 @@ public class OrderService implements IOrderService {
     public List<Order> findOrderWithoutSupplierOffer(String email){
         return orderRepository.findOrderWithoutSupplierOffer(email);
     }
+
+    public void updateWinner(Long winnerId){
+        SupplierOffer supplierOffer = supplierOfferService.getById(winnerId);
+        Order order = supplierOffer.getOrder();
+        List<SupplierOfferDto> supplierOffers= supplierOfferService.offersByOrderId(order.getId());
+        Date date = new Date();
+        if(order.getEndDate().before(date)) {
+            order.setWinnerId(winnerId);
+            orderRepository.save(order);
+            emailService.orderWinner(supplierOffer.getSupplier().getUser().getEmail());
+            supplierOffers.remove(SupplierOfferMapper.mapSupplierOfferToSupplierOfferDto(supplierOffer));
+            for(SupplierOfferDto offer:supplierOffers){
+                if(supplierOffers != null)
+                    emailService.orderNonWinner(offer.getSupplierEmail());
+            }
+            MedicinePharmacy medicinePharmacy = medicinePharmacyService.getByPharmacyAndMedicine(order.getPharmacyAdmin().getPharmacy().getName(),order.getMedicine().getName());
+            if(medicinePharmacy == null){
+                MedicinePharmacy medicinePharmacy1 = new MedicinePharmacy();
+                medicinePharmacy1.setPharmacy(order.getPharmacyAdmin().getPharmacy());
+                medicinePharmacy1.setMedicine(order.getMedicine());
+                medicinePharmacy1.setQuantity(order.getQuantity());
+                medicinePharmacy1.setPrice(order.getPrice());
+                medicinePharmacyService.save(medicinePharmacy1);
+            }else{
+                medicinePharmacy.setQuantity(medicinePharmacy.getQuantity() + order.getQuantity());
+                medicinePharmacyService.save(medicinePharmacy);
+            }
+        }else{
+            throw new InvalidActionException("You can't set winner before expiration date");
+        }
+    }
+
+    public List<Order> getAllByAdmin(String adminEmail){
+        List<Order> orders= orderRepository.findAll();
+        PharmacyAdmin pharmacyAdmin = pharmacyAdminService.findPharmacyAdminByEmail(adminEmail);
+        List<Order> deleteList = new ArrayList<>();
+        for(Order o:orders){
+            if(!o.getPharmacyAdmin().getPharmacy().equals(pharmacyAdmin.getPharmacy()))
+                deleteList.add(o);
+        }
+        orders.removeAll(deleteList);
+        if(orders.isEmpty())
+            throw new NotFoundException("There is no any order.");
+        return orders;
+    }
+
+
 
 }
